requirements.txt
PyMySQL==1.1.0
bcrypt==4.1.1
python-dotenv==1.0.0
pytest==7.4.3
